fetch status of all bugs, not just new ones
this means we re-write the fetch method to filter out ignored bugs, not
just ones we know about; we also have to make the update-y code handle
the case where there's already a bug with that id, and just swap in the
hash instead of replacing it.

We now need to handle if a status change occurs as part of that hash update
(e.g. a bug is now "done", or something has happened to wake up a sleeping
bug)