add a "autocomplete" button
the autocomplete button checks existing bugs and sees whether the
nextaction text you have typed corresponds to an existing nextaction
text. If so, it inserts the matching nextaction text into the field, but
sets the text selection such that if you continued to type, you would
erase the extra text.

Next, this action needs to be tied to typing in the field rather than
pressing the button (and the button should be removed).