initial notebook stuff
We replace the area of the main UI which features the treeview list of triaged
bugs with a notebook.

One page of the notebook contains the treeview that was there before

two other pages contain labels

	one label should be displayed when the bug database is empty and
	instructs the user to click fetch

	another should be displayed when we have a bug db but no triaged
	bugs and instructs the user to triage some bugs

the intention is the code should detect the situations in which to display
each notebook page automatically. then we hide the notebook labels.